Switches to using the correct mechanism for rounding on cost input.
The code had 3 mechanisms for setting the `minValue` and `maxValue` for the input cost, with 2 commented out.  This change removes the incorrect two and uncomments the correct one.

Note: The code for one of the solutions was deleted, but it will be retained in git history if we need it back for something later.