Lots of styling work.
This PR started as a simple fix of just swapping the position of leverage size/multiplier and hiding the exchange cost section.  However, in the process I ended up doing quite a bit of touch-up work on the stylesheets to make them use `grid` and `flex` containers everywhere.

I also consolidated a lot of the styles and deleted a bunch of dead styles to help maintainability.  Hopefully changes going forward will go smoother.

Along the way, I also got the site working in Edge, though it isn't as pretty as Firefox/Chrome since it doesn't support clip-path which is used for the triangle slider.

The site now also works well down to 260px wide, though the chart touch still doesn't function.