Rate limits exchange cost fetches and shows spinner while waiting.
A user could cause great harm to the backend server by just dragging the leverage slider back and forth which results rapid (every tick of the event loop) requests to the remote node in the form of an `eth_call`.  This is now rate limited such that the user needs to not change the leverage multiplier or amount for 500ms before the request is made, at which point it will make the request on behalf.

In order to maintain a quality user experience, a spinner has been added to let the user know that they are waiting on data from the server.  This also prevents them from clicking through to the end without reviewing the exchange costs.

Also fixes a minor bug where the leverage amount buttons were selectable.