Updates client library to latest and adds yarn integrity hashes.
The client library update resolves a problem where if exchange costs were low enough, the UI would break.  It achieves this always suggesting a minimum exchange fee of 1% of the loan amount.  The idea here is that is probably low enough that the user doesn't care, and they'll get a refund on any unspent amount anyway.  It also reduces the chance the user will have a failed transaction, and those chances ego up the lower the exchange costs are.

I also upgraded my yarn version so we have integrity checks on all of the files now.