Adds favicons and switches to hash routing.
Hash routing works much better with IPFS than path routing since it is always interpreted client side.

Favicons for now are just Ethereum icons.  Better than nothing or the Angular icon.

Also put in an attempted fix for a bug where the cost slider would start in the middle instead of on the left.