Updates contract address and fixes bug for real this time.
I thought I had fixed the bug where going back after landing on the error/success page would result in needing to wait for a full polling cycle before values showed up, but I was wrong (I even tested it... I thought!).  The `main-page` is torn down and reconstructed when you navigate to success/fail pages, so the shareReplay needed to be moved up the stack so it would be retained when the page is rebuilt.